+++ src/test/java/net/sf/jabref/logic/util/io/FileNameCleanerTest.java
@@ -1,4 +1,4 @@
-package net.sf.jabref.util;
+package net.sf.jabref.logic.util.io;
 
 import net.sf.jabref.logic.util.io.FileNameCleaner;
 import org.junit.Assert;
